Make injection points into extension points. Also rename them from handlers
